Fixed bug

+++ Block.cpp
@@ -21,7 +21,7 @@ Block::~Block() {
 
 
 void Block::Render(int x, int y, int offset){
-	sprites->Draw(index, x * spriteWidth - offset , y * spriteHeight);
+	sprites->Draw(index, static_cast<float>(x * spriteWidth - offset) , static_cast<float>(y * spriteHeight));
 }
