+++ Clouds.cpp
@@ -1,6 +1,6 @@
 #include "Clouds.h"
 
-Clouds::Clouds(LPCTSTR bitmapPath, Graphics* graphics, int x, int y, float xSpeed)
+Clouds::Clouds(LPCTSTR bitmapPath, Graphics* graphics, float x, float y, float xSpeed)
 	: Animation(bitmapPath, graphics, new GamePoint(x, y), new GameVector(xSpeed, 0), new GameVector(0, 0), false)
 {
